+++ src/AppserverIo/Description/Configuration/PostActivateConfigurationInterface.php
@@ -1,0 +1,41 @@
+<?php
+
+/**
+ * AppserverIo\Description\Configuration\PostActivateConfigurationInterface
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * PHP version 5
+ *
+ * @author    Tim Wagner <tw@appserver.io>
+ * @copyright 2015 TechDivision GmbH <info@appserver.io>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/appserver-io/appserver
+ * @link      http://www.appserver.io
+ */
+
+namespace AppserverIo\Description\Configuration;
+
+/**
+ * Interface for a post activate DTO node implementation.
+ *
+ * @author    Tim Wagner <tw@appserver.io>
+ * @copyright 2015 TechDivision GmbH <info@appserver.io>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/appserver-io/appserver
+ * @link      http://www.appserver.io
+ */
+interface PostActivateConfigurationInterface
+{
+
+    /**
+     * Return's the lifecycle callback methods information.
+     *
+     * @return array The lifecycle callback methods information
+     */
+    public function getLifecycleCallbackMethods();
+}